deletions and insertions in secondary list optimized, needs improvement

+++ server/p.sql
@@ -1,1 +1,1 @@
-delete from users where name == "Elizabeth Jones"; 
+delete from users where uid > 4 and name == "Elizabeth Jones";
